Simplify dotTrace diagnoser by only using JB.SelfApi, dedupe som code by including from shared project
