+++ src/JetBrains/BenchmarkDotNet.Diagnostics.dotMemory/Properties/AssemblyInfo.cs
