+++ src/JetBrains/BenchmarkDotNet.Diagnostics.dotTrace/DotTraceDiagnoserAttribute.cs
