+++ src/JetBrains/BenchmarkDotNet.Diagnostics.dotTrace/Properties/AssemblyInfo.cs
