+++ src/JetBrains/BenchmarkDotNet.JetBrains.Shared/BenchmarkDotNet.JetBrains.Shared.csproj
@@ -1,0 +1,18 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+    <Import Project="..\..\..\build\common.props" />
+    <PropertyGroup>
+        <TargetFrameworks>net6.0;net462;netcoreapp3.1</TargetFrameworks>
+        <NoWarn>$(NoWarn);1591</NoWarn>
+        <AssemblyTitle>BenchmarkDotNet.JetBrains.Shared</AssemblyTitle>
+        <AssemblyName>BenchmarkDotNet.JetBrains.Shared</AssemblyName>
+        <Nullable>enable</Nullable>
+    </PropertyGroup>
+
+    <ItemGroup>
+        <ProjectReference Include="..\..\BenchmarkDotNet\BenchmarkDotNet.csproj" />
+    </ItemGroup>
+    
+    <ItemGroup>
+        <PackageReference Include="JetBrains.Profiler.SelfApi" Version="2.5.2" />
+    </ItemGroup>
+</Project>
