+++ src/JetBrains/BenchmarkDotNet.JetBrains.Shared/Progress.cs
@@ -2,9 +2,10 @@ using System;
 using System.Diagnostics;
 using BenchmarkDotNet.Loggers;
 
-namespace BenchmarkDotNet.Diagnostics.dotMemory
+namespace BenchmarkDotNet.JetBrains.Shared
 {
-    public class Progress : IProgress<double>
+
+    internal sealed class Progress : IProgress<double>
     {
         private static readonly TimeSpan ReportInterval = TimeSpan.FromSeconds(0.1);
